requirements.txt
discord.py
python-dotenv
audioop-lts
requests
pytz
pynacl